Update version and refine status message format
This commit does two main things. Firstly, the version of the project in the build.gradle file is updated from '1.0.1' to '1.0.2'. Secondly, the status messages in the ClearStatusCommand.java and StatusCommand.java files have been slightly adjusted. In the former, the status is stripped of any whitespace before being displayed to the user, while in the latter, the message's line has been simplified for readability. These changes aim to improve the user interface and keep the project version updated.